Working but unstable

+++ Crazyflie_Firmware/src/modules/src/system.c
@@ -184,7 +184,7 @@ void systemTask(void *arg)
   commInit();
   commanderInit();
   //stabilizerInit();
-  xQueue1 = xQueueCreate( 1, sizeof( int* ) );
+  xQueue1 = xQueueCreate( 1, sizeof( float* ) );
   xQueue2 = xQueueCreate( 1, sizeof( int* ) );
   mode_switcherInit();
   controllerInit();
